flwr_tft/pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tft-flower"
version = "0.1.0"
description = "Federated training for dreaminate/tft_module with Flower"
requires-python = ">=3.9"
dependencies = [
  "flwr>=1.19.0",                 # 不拉 ray/otel，更稳
  "pytorch-lightning>=2.2.0",
  "torch",
  "pytorch-forecasting",
  "pyyaml",
  "omegaconf"
]

[tool.setuptools]
# 扁平布局：声明顶层模块，支持 editable 安装
py-modules = ["client_app", "server_app", "task"]

[tool.flwr.app]
publisher = "dreaminate"

[tool.flwr.app.components]
serverapp = "server_app:app"
clientapp = "client_app:app"

[tool.flwr.app.config]
num-server-rounds = 3
fraction-fit = 1.0
fraction-evaluate = 1.0
min-available-clients = 2
local-epochs = 1
batch-size = 256
num-workers = 4

[tool.flwr.federations]
default = "local-simulation"

[tool.flwr.federations.local-simulation]
# 新写法（部分版本识别）
num-supernodes = 4
backend.name = "multiprocessing"
# 兼容老写法（部分版本识别）
options.num-supernodes = 4
options.backend.name = "multiprocessing"
